Move Copy to Clipboard component to own file
